# apps/api/openapi/openapi.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: CaenCamp.s JobBoard API
  description: "Description de l'API Rest des offres d'emplois des CaenCamp.s"
  termsOfService: 'https://jobboard.caen.camp/terms/'
  contact:
    name: JobBoard Support
    url: 'https://jobboard.caen.camp/support'
    email: jobboard@caen.camp
  license:
    name: GNU GPLv3
    url: 'https://github.com/CaenCamp/jobs-caen-camp/blob/master/LICENSE'
servers:
  - url: 'http://localhost:8001'
    description: Local server
security: []
...
paths:
  /api/job-postings:
    get:
      tags:
        - "Offres d'emploi"
      summary: "Liste des offres d'emploi"
      responses:
        '200':
          description: "Liste paginée d'offres d'emploi"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FullJobPosting'
              examples:
                example-1:
                  value:
                    - id: a122edec-5580-4a93-aff7-fc18b41e4c57
                      title: Developpeur Javascript
                      url: 'https://jobs.caen.camp'
                      datePosted: '2010-03-02'
                      employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
                      employmentType: CDD
                      experienceRequirements: "3 ans d'experience sur un projet Javascript"
                      jobStartDate: '2020-05-02'
                      skills: 'JavaScript, Devops, Php, ...'
                      validThrough: '2010-05-05'
                      hiringOrganization:
                        identifier: a122edec-5580-4a93-aff7-fc18b41e4c57
                        name: Incaya
                        image: 'https://www.incaya.fr/static/logo-incaya.svg'
                        url: 'https://www.incaya.fr/'
                        address:
                          addressCountry: France
                          addressLocality: Caen
                          postalCode: '14000'
          headers:
            X-Total-Count:
                $ref: '#/components/headers/X-Total-Count'
            Link:
                $ref: '#/components/headers/Link'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      operationId: getJobPostings
      description: "Obtenir une liste paginée d'offres d'emploi\n\nCette liste est filtrable par filters={ key: value, key2: value ...} :\n* title\n* skills\n* employmentType\n* datePosted_before\n* datePosted_after\n* jobStartDate_before\n* jobStartDate_after\n* validThrough_before\n* validThrough_after\n* hiringOrganizationName\n* hiringOrganizationPostalCode\n* hiringOrganizationAddressLocality\n* hiringOrganizationAddressCountry\n\nCette liste est triable par sort=[key, sortDirection_ASC_DESC] : \n* datePosted\n* title\n* jobStartDate\n* validThrough\n* hiringOrganizationName\n* hiringOrganizationIdentifier\n* hiringOrganizationPostalCode\n* hiringOrganizationAddressLocality\n* hiringOrganizationAddressCountry\n\nLa pagination est controlable via le paramètre de requête pagination=[nombreDeRésultatsParPage, numeroDePage]"
      parameters:
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/PaginationCurrentPage'
        - $ref: '#/components/parameters/PaginationPerPage'
    post:
      tags:
        - "Offres d'emploi"
      summary: "Création d'une offre d'emploi"
      operationId: createJobPosting
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/JobPosting'
                - type: object
                  properties:
                    hiringOrganizationId:
                      type: string
                      format: uuid
                      example: 9b24d771-1069-40b1-b6be-3f3e53fdadd0
                  required:
                    - hiringOrganizationId
            examples:
              example-1:
                value:
                  title: Developpeur Javascript
                  url: 'https://jobs.caen.camp'
                  datePosted: '2010-03-02'
                  employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
                  employmentType: CDD
                  experienceRequirements: "3 ans d'experience sur un projet Javascript"
                  jobStartDate: '2020-05-02'
                  skills: 'JavaScript, Devops, Php, ...'
                  validThrough: '2010-05-05'
                  hiringOrganizationId: a122edec-5580-4a93-aff7-fc18b41e4c57
        description: "Données permettant de créer une offre d'emploi"
      description: "Ajouter une offre d'emploi au JobBoard"
      responses:
        '200':
          description: "Retourne L'offre d'emploi créée"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobPosting'
              examples:
                example-1:
                  value:
                    id: a122edec-5580-4a93-aff7-fc18b41e4c57
                    title: Developpeur Javascript
                    url: 'https://jobs.caen.camp'
                    datePosted: '2010-03-02'
                    employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
                    employmentType: CDD
                    experienceRequirements: "3 ans d'experience sur un projet Javascript"
                    jobStartDate: '2020-05-02'
                    skills: 'JavaScript, Devops, Php, ...'
                    validThrough: '2010-05-05'
                    hiringOrganizationId: a122edec-5580-4a93-aff7-fc18b41e4c57
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      parameters: []
    parameters: []
  '/api/job-postings/{identifier}':
    get:
      tags:
        - "Offres d'emploi"
      summary: Trouver une offre par son ID
      description: "Renvoi une offre d'emploi"
      operationId: getJobById
      responses:
        '200':
          description: "L'offre d'emploi demandée"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullJobPosting'
              examples:
                example-1:
                  value:
                    id: a122edec-5580-4a93-aff7-fc18b41e4c57
                    title: Developpeur Javascript
                    url: 'https://jobs.caen.camp'
                    datePosted: '2010-03-02'
                    employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
                    employmentType: CDD
                    experienceRequirements: "3 ans d'experience sur un projet Javascript"
                    jobStartDate: '2020-05-02'
                    skills: 'JavaScript, Devops, Php, ...'
                    validThrough: '2010-05-05'
                    hiringOrganization:
                      identifier: a122edec-5580-4a93-aff7-fc18b41e4c57
                      name: Incaya
                      image: 'https://www.incaya.fr/static/logo-incaya.svg'
                      url: 'https://www.incaya.fr/'
                      address:
                        addressCountry: France
                        addressLocality: Caen
                        postalCode: '14000'
        '400':
          description: ID fourni invalide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '404':
          description: "Offre d'emploi non trouvée"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
    parameters:
      - $ref: '#/components/parameters/UUID'
    put:
      tags:
        - "Offres d'emploi"
      summary: "Mettre à jour une offre d'emploi"
      operationId: updateJobPosting
      responses:
        '200':
          description: "Offre d'emploi mise à jour"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullJobPosting'
              examples:
                example-1:
                  value:
                    id: a122edec-5580-4a93-aff7-fc18b41e4c57
                    title: Developpeur Javascript
                    url: 'https://jobs.caen.camp'
                    datePosted: '2010-03-02'
                    employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
                    employmentType: CDD
                    experienceRequirements: "3 ans d'experience sur un projet Javascript"
                    jobStartDate: '2020-05-02'
                    skills: 'JavaScript, Devops, Php, ...'
                    validThrough: '2010-05-05'
                    hiringOrganization:
                      identifier: a122edec-5580-4a93-aff7-fc18b41e4c57
                      name: Incaya
                      image: 'https://www.incaya.fr/static/logo-incaya.svg'
                      url: 'https://www.incaya.fr/'
                      address:
                        addressCountry: France
                        addressLocality: Caen
                        postalCode: '14000'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      description: "Editer une offre d'emploi"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobPosting'
            examples:
              example-1:
                value:
                  title: Developpeur Javascript
                  url: 'https://jobs.caen.camp'
                  datePosted: '2010-03-02'
                  employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
                  employmentType: CDD
                  experienceRequirements: "3 ans d'experience sur un projet Javascript"
                  jobStartDate: '2020-05-02'
                  skills: 'JavaScript, Devops, Php, ...'
                  validThrough: '2010-05-05'
                  hiringOrganizationId: a122edec-5580-4a93-aff7-fc18b41e4c57
        description: "Données permettant de mettre à jour une offre d'emploi"
    delete:
      tags:
        - "Offres d'emploi"
      summary: "Supprimer une offre d'emploi"
      operationId: deleteJobPosting
      responses:
        '200':
          description: "Offre d'emploi supprimée"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                example-1:
                  value:
                    id: a122edec-5580-4a93-aff7-fc18b41e4c57
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorModel'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtendedErrorModel'
      description: "Supprimer une offre d'emploi"
...
components:
  securitySchemes: {}
  headers:
    X-Total-Count:
        description: "Return information about total items to paginate"
        schema:
            type: integer
        example: 42
        required: false
    Link:
        description: "Return information about pagination actions"
        schema:
            type: string
        example: |
            </resources?currentPage=3&perPage=10>; rel="self",
            </resources?currentPage=1&perPage=10>; rel="first",
            </resources?currentPage=2&perPage=10>; rel="prev",
            </resources?currentPage=4&perPage=10>; rel="next",
            </resources?currentPage=5&perPage=10>; rel="last",
        required: false
  schemas:
    JobPosting:
      description: "Une offre d'emploi"
      type: object
      x-examples:
        example-devReact:
          title: Developpeur Javascript
          url: 'https://jobs.caen.camp'
          datePosted: '2010-03-02'
          employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
          employmentType: CDD
          experienceRequirements: "3 ans d'experience sur un projet Javascript"
          jobStartDate: '2020-05-02'
          skills: 'JavaScript, Devops, Php, ...'
          validThrough: '2010-05-05'
      properties:
        title:
          type: string
          example: Developpeur Cobol
        url:
          type: string
          format: uri
          minLength: 10
        datePosted:
          type: string
          format: date
          example: '2010-03-02'
        employerOverview:
          type: string
          example: "Au sein d'une équipe DSI composée de juniors et de séniors"
        employmentType:
          type: string
          enum:
            - CDI
            - CDD
            - Alternance
            - Autre
          example: CDD
        experienceRequirements:
          type: string
          example: "3 ans d'experience sur un projet Javascript"
          minLength: 20
        jobStartDate:
          type: string
          format: date
          nullable: true
          example: '2020-05-02'
        skills:
          type: string
          example: 'JavaScript, Devops, Php, ...'
        validThrough:
          type: string
          format: date
          nullable: true
          example: '2010-05-05'
      required:
        - datePosted
        - employerOverview
        - employmentType
        - experienceRequirements
        - skills
    Organization:
      description: Une entreprise liée au développement informatique.
      type: object
      x-examples:
        example-hiringOrganization:
          name: Incaya
          description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
          image: 'https://www.incaya.fr/static/logo-incaya.svg'
          email: contact@organization.org
          url: 'https://www.incaya.fr/'
          address:
            addressCountry: France
            addressLocality: Caen
            postalCode: '14000'
            streetAddress: '5, place de la Répulique'
          contactPoints:
            - identifier: abeb9875-2191-4097-8249-5b8defc42a3e
              email: job@organization.org
              telephone: '0606060606'
              name: 'John Do, CTO'
              contactType: "Offres d'emploi"
      properties:
        name:
          type: string
          example: Incaya
          minLength: 2
          maxLength: 150
        description:
          type: string
          example: Coopérative de services du numérique
        url:
          type: string
          format: uri
          example: 'https://www.incaya.fr/'
          minLength: 10
          pattern: "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)"
        email:
          type: string
          format: email
          example: email@organization.org
        image:
          type: string
          format: uri
          example: 'https://www.incaya.fr/static/logo-incaya.svg'
          minLength: 10
          pattern: "https?:\\/\\/(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)"
        address:
          anyOf:
            - type: string
            - $ref: '#/components/schemas/PostalAddress'
        contactPoints:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/FullContactPoint'
      required:
        - name
        - url
    PostalAddress:
      title: PostalAddress
      type: object
      properties:
        addressCountry:
          type: string
          example: France
          minLength: 2
        addressLocality:
          type: string
          example: Caen
          minLength: 2
        postalCode:
          type: string
          example: '14000'
          minLength: 5
          maxLength: 5
        streetAddress:
          type: string
          example: '5, place de la Répulique'
      required:
        - addressLocality
        - postalCode
      description: Une adresse postale.
      x-examples:
        example-Caen:
          addressCountry: France
          addressLocality: Caen
          postalCode: '14000'
          streetAddress: '5, place de la Répulique'
    ContactPoint:
      title: ContactPoint
      type: object
      x-examples:
        example-job:
          email: job@organization.org
          telephone: '0606060606'
          name: 'John Do, CTO'
          contactType: "Offres d'emploi"
      description: "Une personne à contacter au sein de l'organisation"
      properties:
        email:
          type: string
          format: email
          example: job@organization.org
          minLength: 7
        telephone:
          type: string
          nullable: true
        name:
          type: string
        contactType:
          type: string
          example: "Offres d'emploi"
          nullable: true
      required:
        - email
        - name
    ExtendedErrorModel:
      allOf:
        - $ref: '#/components/schemas/ErrorModel'
        - type: object
          properties:
            rootCause:
              type: string
              example: Connexion to db failed
          required:
            - rootCause
      description: "Une message d'erreur avec la stack d'erreur"
      x-examples:
        example-1: {}
    ErrorModel:
      type: object
      description: "Un message d'erreur"
      x-examples:
        example-404:
          code: 404
          message: "Cette offre d'emploi n'existe pas."
      properties:
        message:
          type: string
          example: "Cette offre d'emploi n'existe pas"
          minLength: 10
        code:
          type: integer
          format: int32
          minimum: 200
          maximum: 500
      required:
        - message
        - code
    Identifier:
      title: Identifier
      type: object
      description: "L'identifiant unique des objets persistés dans le JobBoard"
      properties:
        identifier:
          type: string
          format: uuid
          example: a122edec-5580-4a93-aff7-fc18b41e4c57
    FullJobPosting:
      title: FullJobPosting
      description: "Une offre d'emploi complète (avec identifiant unique et entreprise liée)"
      allOf:
        - $ref: '#/components/schemas/Identifier'
        - $ref: '#/components/schemas/JobPosting'
        - type: object
          properties:
            hiringOrganization:
              $ref: '#/components/schemas/FullOrganization'
      x-examples:
        example-1:
          title: Developpeur Javascript
          hiringOrganization:
            identifier: a122edec-5580-4a93-aff7-fc18b41e4c57
            name: Incaya
            description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
            image: 'https://www.incaya.fr/static/logo-incaya.svg'
            email: contact@organization.org
            url: 'https://www.incaya.fr/'
            address:
              addressCountry: France
              addressLocality: Caen
              postalCode: '14000'
              streetAddress: '5, place de la Répulique'
            contactPoints:
              - email: job@organization.org
                telephone: '0606060606'
                name: 'John Do, CTO'
                contactType: "Offres d'emploi"
          url: 'https://jobs.caen.camp'
          datePosted: '2010-03-02'
          employerOverview: "Au sein d'une équipe DSI composée de juniors et de séniors"
          employmentType: CDD
          experienceRequirements: "3 ans d'experience sur un projet Javascript"
          jobStartDate: '2020-05-02'
          skills: 'JavaScript, Devops, Php, ...'
          validThrough: '2010-05-05'
    FullOrganization:
      title: FullOrganization
      description: Une entreprise complete avec identifiant unique
      allOf:
        - $ref: '#/components/schemas/Identifier'
        - $ref: '#/components/schemas/Organization'
      x-examples:
        example-1:
          identifier: a122edec-5580-4a93-aff7-fc18b41e4c57
          name: Incaya
          description: "INCAYA, société coopérative de services du numérique. Créons ensemble des outils qui ont du sens ! Collectivités, associations, entreprises : nous concevons avec vous des outils numériques qui transforment l'expérience de vos utilisateurs et usagers."
          image: 'https://www.incaya.fr/static/logo-incaya.svg'
          email: contact@organization.org
          url: 'https://www.incaya.fr/'
          address:
            addressCountry: France
            addressLocality: Caen
            postalCode: '14000'
            streetAddress: '5, place de la Répulique'
          contactPoints:
            - email: job@organization.org
              telephone: '0606060606'
              name: 'John Do, CTO'
              contactType: "Offres d'emploi"
    FullContactPoint:
      title: FullContactPoint
      allOf:
        - $ref: '#/components/schemas/Identifier'
        - $ref: '#/components/schemas/ContactPoint'
      x-examples:
        example-1:
          identifier: abeb9875-2191-4097-8249-5b8defc42a3e
          email: job@organization.org
          telephone: '0606060606'
          name: 'John Do, CTO'
          contactType: "Offres d'emploi"
...
  parameters:
    Filter:
      name: filters
      in: query
      description: "Filtres appliquables à la requête. C'est un objet json stringifié, avec les filtres en clé/valeur séparé par des virgules."
      required: false
      explode: true
      schema:
        type: string
    Sort:
      name: sort
      in: query
      description: "Le tri applicable à la liste. C'est un tableau stringifié de la forme [sortProp, sortDirection]"
      required: false
      explode: false
      schema:
        type: array
        maxItems: 2
        minItems: 2
        items:
          type: string
    PaginationCurrentPage:
      name: currentPage
      in: query
      description: "Le paramètre de pagination pour définir la page courrante."
      required: false
      schema:
        type: integer
        example: 1
    PaginationPerPage:
      name: perPage
      in: query
      description: "Le paramètre de pagination pour définir le nombre d'éléments par page."
      required: false
      schema:
        type: integer
        example: 10
    UUID:
      name: identifier
      in: path
      description: "Identifiant unique d'un objet de forme uuid"
      required: true
      schema:
        type: string
        format: uuid
      example: 9a6c8995-df54-446c-a5b8-71532c304751